worker/requirements.txt
pandas
mysql-connector-python
boto3